pyproject.toml
[project]
name = "launchboard"
version = "0.2.0"
description = "AI-powered job search pipeline — scrape, score, optimize, and track job applications"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10,<3.13"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]
dependencies = [
    "python-jobspy>=1.1.75",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "PyPDF2>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "ddgs>=9.0.0",
    "keyring>=25.0.0",
    "cryptography>=42.0.0",
]

[project.scripts]
launchboard = "job_finder.main:run"
launchboard-setup = "job_finder.setup:run_setup"
job-finder = "job_finder.main:run"
search-jobs = "job_finder.main:search"
score-jobs = "job_finder.main:score"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/bteh/launchboard"
Repository = "https://github.com/bteh/launchboard"
Issues = "https://github.com/bteh/launchboard/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/job_finder"]
